Only author of post can see options

+++ socialapp/src/components/Posts/PostCard.jsx
@@ -1,25 +1,28 @@
 import React, { useState, useEffect, useContext } from 'react';
 import ReactMarkdown from 'react-markdown';
-import axios from 'axios';
 import { Dropdown } from 'react-bootstrap';
 import { BiDotsVerticalRounded } from "react-icons/bi";
 import Card from 'react-bootstrap/Card';
 import AuthContext from '../../context/AuthContext';
 import "./PostCard.css";
+import useAxios from "../../utils/useAxios";
+
+
 
 export default function PostCard(props) {
   const user_id = localStorage.getItem("user_id");
   const [isOwner, setIsOwner] = useState(true);
   const { baseURL } = useContext(AuthContext);      // our api url http://127.0.0.1/service
   const { authTokens } = useContext(AuthContext);
+  const api = useAxios();
   
   useEffect (() => {
 
   }, []);
 
   const deletePost = (uuid) => {
-      axios
-        .delete(`${baseURL}/authors/${user_id}/posts/${uuid}`, { headers: { 'Authorization': `Bearer ${authTokens.access}` } })
+      api
+        .delete(`${baseURL}/authors/${user_id}/posts/${uuid}`)
         .then((response) => {
           console.log(response.data);
           window.location.reload(true);
@@ -30,22 +33,33 @@ export default function PostCard(props) {
         });
   };
 
+  // only render options if the user viewing it is the author of it
+  function PostOptions () {
+    if (user_id === props.post.author.uuid) {
+      return (
+        <div className="options">
+        <Dropdown>
+          <Dropdown.Toggle id="dropdown-basic">
+            <BiDotsVerticalRounded />
+          </Dropdown.Toggle>
+          <Dropdown.Menu>
+            {isOwner && <Dropdown.Item onClick={() => deletePost(props.post.uuid)}>Delete Post</Dropdown.Item>}
+          </Dropdown.Menu>  
+        </Dropdown>
+      </div>
+      );
+    } else {
+      return (<></>);
+    }
+  }
+
   return (
     <Card className="post-card"style={{ width: '30rem', whiteSpace: "nowrap"}}>
       <Card.Header>
-        <div style={{display: "inline-block"}} className="profilePicPage">
+        <div className="profilePicPost">
           <img id="profilePicPage" src={props.post.author.profileImage} alt="profilePic"/>
         </div>
-        <div className="options">
-          <Dropdown>
-            <Dropdown.Toggle id="dropdown-basic">
-              <BiDotsVerticalRounded />
-            </Dropdown.Toggle>
-            <Dropdown.Menu>
-              {isOwner && <Dropdown.Item onClick={() => deletePost(props.post.uuid)}>Delete Post</Dropdown.Item>}
-            </Dropdown.Menu>  
-          </Dropdown>
-        </div>
+        <PostOptions />
       </Card.Header>
       <Card.Img variant="top" src="" />
       <Card.Body>
